+++ src/providers/aliyun/mock_tests.rs
@@ -7,63 +7,54 @@ fn basic_hostname() {
     let ep = "/latest/meta-data/hostname";
     let hostname = "test-hostname";
 
+    let mut server = mockito::Server::new();
     let mut provider = aliyun::AliyunProvider::try_new().unwrap();
-    provider.client = provider
-        .client
-        .max_retries(0)
-        .mock_base_url(mockito::server_url());
+    provider.client = provider.client.max_retries(0).mock_base_url(server.url());
 
-    {
-        let _m503 = mockito::mock("GET", ep).with_status(503).create();
-        provider.hostname().unwrap_err();
-    }
+    server.mock("GET", ep).with_status(503).create();
+    provider.hostname().unwrap_err();
 
-    {
-        let _m200 = mockito::mock("GET", ep)
-            .with_status(200)
-            .with_body(hostname)
-            .create();
-        let v = provider.hostname().unwrap();
-        assert_eq!(v, Some(hostname.to_string()));
-    }
+    server
+        .mock("GET", ep)
+        .with_status(200)
+        .with_body(hostname)
+        .create();
+    let v = provider.hostname().unwrap();
+    assert_eq!(v, Some(hostname.to_string()));
 
-    {
-        let _m404 = mockito::mock("GET", ep).with_status(404).create();
-        let v = provider.hostname().unwrap();
-        assert_eq!(v, None);
-    }
+    server.mock("GET", ep).with_status(404).create();
+    let v = provider.hostname().unwrap();
+    assert_eq!(v, None);
 
-    {
-        let _m200_empty = mockito::mock("GET", ep)
-            .with_status(200)
-            .with_body("")
-            .create();
-        let v = provider.hostname().unwrap();
-        assert_eq!(v, None);
-    }
+    server
+        .mock("GET", ep)
+        .with_status(200)
+        .with_body("")
+        .create();
+    let v = provider.hostname().unwrap();
+    assert_eq!(v, None);
 
-    mockito::reset();
+    server.reset();
     provider.hostname().unwrap_err();
 }
 
 #[test]
 fn basic_pubkeys() {
+    let mut server = mockito::Server::new();
     let mut provider = aliyun::AliyunProvider::try_new().unwrap();
-    provider.client = provider
-        .client
-        .max_retries(0)
-        .mock_base_url(mockito::server_url());
+    provider.client = provider.client.max_retries(0).mock_base_url(server.url());
 
     // Setup two entries with identical content, in order to test de-dup.
-    let _m_keys = mockito::mock("GET", "/latest/meta-data/public-keys/")
+    server
+        .mock("GET", "/latest/meta-data/public-keys/")
         .with_status(200)
         .with_body("0/\ntest/\n")
         .create();
-    let _m_key0 = mockito::mock("GET", "/latest/meta-data/public-keys/0/openssh-key")
+    server.mock("GET", "/latest/meta-data/public-keys/0/openssh-key")
         .with_status(200)
         .with_body("ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIIadOopfaOOAdFWRkCoOimvDyOftqphtnIeiECJuhkdq test-comment")
         .create();
-    let _m_keytest = mockito::mock("GET", "/latest/meta-data/public-keys/test/openssh-key")
+    server.mock("GET", "/latest/meta-data/public-keys/test/openssh-key")
         .with_status(200)
         .with_body("ssh-ed25519 AAAAC3NzaC1lZDI1NTE5AAAAIIadOopfaOOAdFWRkCoOimvDyOftqphtnIeiECJuhkdq test-comment")
         .create();
@@ -75,7 +66,7 @@ fn basic_pubkeys() {
     assert_eq!(keys[0].options, None);
     assert_eq!(keys[0].comment, Some("test-comment".to_string()));
 
-    mockito::reset();
+    server.reset();
     provider.ssh_keys().unwrap_err();
 }
 
@@ -104,13 +95,13 @@ fn basic_attributes() {
         "/latest/meta-data/vpc-id" => vpc_id,
         "/latest/meta-data/zone-id" => zone_id,
     };
-    let mut mocks = Vec::with_capacity(endpoints.len());
+    let mut server = mockito::Server::new();
     for (endpoint, body) in endpoints {
-        let m = mockito::mock("GET", endpoint)
+        server
+            .mock("GET", endpoint)
             .with_status(200)
             .with_body(body)
             .create();
-        mocks.push(m);
     }
 
     let attributes = maplit::hashmap! {
@@ -130,12 +121,12 @@ fn basic_attributes() {
         .unwrap()
         .max_retries(0)
         .return_on_404(true)
-        .mock_base_url(mockito::server_url());
+        .mock_base_url(server.url());
     let provider = aliyun::AliyunProvider { client };
 
     let v = provider.attributes().unwrap();
     assert_eq!(v, attributes);
 
-    mockito::reset();
+    server.reset();
     provider.attributes().unwrap_err();
 }