+++ src/providers/gcp/mock_tests.rs
@@ -7,33 +7,26 @@ fn basic_hostname() {
     let ep = "/computeMetadata/v1/instance/hostname";
     let hostname = "test-hostname";
 
+    let mut server = mockito::Server::new();
     let mut provider = gcp::GcpProvider::try_new().unwrap();
-    provider.client = provider
-        .client
-        .max_retries(0)
-        .mock_base_url(mockito::server_url());
+    provider.client = provider.client.max_retries(0).mock_base_url(server.url());
 
-    {
-        let _m503 = mockito::mock("GET", ep).with_status(503).create();
-        provider.hostname().unwrap_err();
-    }
+    server.mock("GET", ep).with_status(503).create();
+    provider.hostname().unwrap_err();
 
-    {
-        let _m200 = mockito::mock("GET", ep)
-            .with_status(200)
-            .with_body(hostname)
-            .create();
-        let v = provider.hostname().unwrap();
-        assert_eq!(v, Some(hostname.to_string()));
-    }
+    server
+        .mock("GET", ep)
+        .with_status(200)
+        .with_body(hostname)
+        .create();
+    let v = provider.hostname().unwrap();
+    assert_eq!(v, Some(hostname.to_string()));
 
-    {
-        let _m404 = mockito::mock("GET", ep).with_status(404).create();
-        let v = provider.hostname().unwrap();
-        assert_eq!(v, None);
-    }
+    server.mock("GET", ep).with_status(404).create();
+    let v = provider.hostname().unwrap();
+    assert_eq!(v, None);
 
-    mockito::reset();
+    server.reset();
     provider.hostname().unwrap_err();
 }
 
@@ -50,13 +43,13 @@ fn basic_attributes() {
         "/computeMetadata/v1/instance/network-interfaces/0/ip" => ip_local,
         "/computeMetadata/v1/instance/machine-type" => machine_type,
     };
-    let mut mocks = Vec::with_capacity(endpoints.len());
+    let mut server = mockito::Server::new();
     for (endpoint, body) in endpoints {
-        let m = mockito::mock("GET", endpoint)
+        server
+            .mock("GET", endpoint)
             .with_status(200)
             .with_body(body)
             .create();
-        mocks.push(m);
     }
 
     let attributes = maplit::hashmap! {
@@ -70,12 +63,12 @@ fn basic_attributes() {
         .unwrap()
         .max_retries(0)
         .return_on_404(true)
-        .mock_base_url(mockito::server_url());
+        .mock_base_url(server.url());
     let provider = gcp::GcpProvider { client };
 
     let v = provider.attributes().unwrap();
     assert_eq!(v, attributes);
 
-    mockito::reset();
+    server.reset();
     provider.attributes().unwrap_err();
 }