+++ src/providers/microsoft/azurestack/mock_tests.rs
@@ -67,7 +67,7 @@ static GOALSTATE_BODY_NO_CERTS: &str = r#"<?xml version="1.0" encoding="utf-8"?>
 </GoalState>
 "#;
 
-fn mock_fab_version() -> mockito::Mock {
+fn mock_fab_version(server: &mut mockito::Server) -> mockito::Mock {
     let fab_version = "/?comp=versions";
     let ver_body = r#"<?xml version="1.0" encoding="utf-8"?>
 <Versions>
@@ -88,13 +88,14 @@ fn mock_fab_version() -> mockito::Mock {
   </Supported>
 </Versions>"#;
 
-    mockito::mock("GET", fab_version)
+    server
+        .mock("GET", fab_version)
         .with_body(ver_body)
         .with_status(200)
         .create()
 }
 
-fn mock_goalstate(with_certificates: bool) -> mockito::Mock {
+fn mock_goalstate(server: &mut mockito::Server, with_certificates: bool) -> mockito::Mock {
     let fab_goalstate = "/machine/?comp=goalstate";
 
     let gs_body = if with_certificates {
@@ -103,7 +104,8 @@ fn mock_goalstate(with_certificates: bool) -> mockito::Mock {
         GOALSTATE_BODY_NO_CERTS
     };
 
-    mockito::mock("GET", fab_goalstate)
+    server
+        .mock("GET", fab_goalstate)
         .with_body(gs_body)
         .with_status(200)
         .create()
@@ -111,11 +113,13 @@ fn mock_goalstate(with_certificates: bool) -> mockito::Mock {
 
 #[test]
 fn test_boot_checkin() {
-    let m_version = mock_fab_version();
-    let m_goalstate = mock_goalstate(true);
+    let mut server = mockito::Server::new();
+    let m_version = mock_fab_version(&mut server);
+    let m_goalstate = mock_goalstate(&mut server, true);
 
     let fab_health = "/machine/?comp=health";
-    let m_health = mockito::mock("POST", fab_health)
+    let m_health = server
+        .mock("POST", fab_health)
         .match_header("content-type", Matcher::Regex("text/xml".to_string()))
         .match_header("x-ms-version", Matcher::Regex("2012-11-30".to_string()))
         .match_body(Matcher::Regex("<State>Ready</State>".to_string()))
@@ -127,7 +131,7 @@ fn test_boot_checkin() {
 
     let client = retry::Client::try_new()
         .unwrap()
-        .mock_base_url(mockito::server_url());
+        .mock_base_url(server.url());
     let provider = azurestack::AzureStack::with_client(Some(client)).unwrap();
     let r = provider.boot_checkin();
 
@@ -136,7 +140,7 @@ fn test_boot_checkin() {
     m_health.assert();
     r.unwrap();
 
-    mockito::reset();
+    server.reset();
 
     // Check error logic, but fail fast without re-trying.
     let client = retry::Client::try_new().unwrap().max_retries(0);
@@ -145,12 +149,14 @@ fn test_boot_checkin() {
 
 #[test]
 fn test_identity() {
-    let m_version = mock_fab_version();
+    let mut server = mockito::Server::new();
+    let m_version = mock_fab_version(&mut server);
 
     let testname = "testName"; // TODO clean up response composition
     let json_response = r#"{"subscriptionId":"testID","vmName":"testName"}"#;
     let endpoint = "/Microsoft.Compute/identity?api-version=2019-03-11";
-    let m_identity = mockito::mock("GET", endpoint)
+    let m_identity = server
+        .mock("GET", endpoint)
         .match_header("Metadata", "true")
         .with_body(json_response)
         .with_status(200)
@@ -158,7 +164,7 @@ fn test_identity() {
 
     let client = retry::Client::try_new()
         .unwrap()
-        .mock_base_url(mockito::server_url());
+        .mock_base_url(server.url());
     let provider = azurestack::AzureStack::with_client(Some(client)).unwrap();
     let r = provider.hostname().unwrap();
 
@@ -168,7 +174,7 @@ fn test_identity() {
     let hostname = r.unwrap();
     assert_eq!(hostname, testname);
 
-    mockito::reset();
+    server.reset();
 
     // Check error logic, but fail fast without re-trying.
     let client = retry::Client::try_new().unwrap().max_retries(0);
@@ -177,12 +183,14 @@ fn test_identity() {
 
 #[test]
 fn test_hostname() {
-    let m_version = mock_fab_version();
+    let mut server = mockito::Server::new();
+    let m_version = mock_fab_version(&mut server);
 
     let testname = "testName"; // TODO clean up response composition
     let json_response = r#"{"subscriptionId":"testID","vmName":"testName"}"#;
     let endpoint = "/Microsoft.Compute/identity?api-version=2019-03-11";
-    let m_identity = mockito::mock("GET", endpoint)
+    let m_identity = server
+        .mock("GET", endpoint)
         .match_header("Metadata", "true")
         .with_body(json_response)
         .with_status(200)
@@ -190,7 +198,7 @@ fn test_hostname() {
 
     let client = retry::Client::try_new()
         .unwrap()
-        .mock_base_url(mockito::server_url());
+        .mock_base_url(server.url());
     let provider = azurestack::AzureStack::with_client(Some(client)).unwrap();
     let r = provider.hostname().unwrap();
 
@@ -200,7 +208,7 @@ fn test_hostname() {
     let hostname = r.unwrap();
     assert_eq!(hostname, testname);
 
-    mockito::reset();
+    server.reset();
 
     // Check error logic, but fail fast without re-trying.
     let client = retry::Client::try_new().unwrap().max_retries(0);
@@ -209,12 +217,13 @@ fn test_hostname() {
 
 #[test]
 fn test_goalstate_certs() {
-    let m_version = mock_fab_version();
-    let m_goalstate = mock_goalstate(true);
+    let mut server = mockito::Server::new();
+    let m_version = mock_fab_version(&mut server);
+    let m_goalstate = mock_goalstate(&mut server, true);
 
     let client = retry::Client::try_new()
         .unwrap()
-        .mock_base_url(mockito::server_url());
+        .mock_base_url(server.url());
     let provider = azurestack::AzureStack::with_client(Some(client)).unwrap();
     let goalstate = provider.fetch_goalstate().unwrap();
 
@@ -225,17 +234,18 @@ fn test_goalstate_certs() {
     let certs_url = reqwest::Url::parse(&ep).unwrap();
     assert_eq!(certs_url.scheme(), "http");
 
-    mockito::reset();
+    server.reset();
 }
 
 #[test]
 fn test_goalstate_no_certs() {
-    let m_version = mock_fab_version();
-    let m_goalstate = mock_goalstate(false);
+    let mut server = mockito::Server::new();
+    let m_version = mock_fab_version(&mut server);
+    let m_goalstate = mock_goalstate(&mut server, false);
 
     let client = retry::Client::try_new()
         .unwrap()
-        .mock_base_url(mockito::server_url());
+        .mock_base_url(server.url());
     let provider = azurestack::AzureStack::with_client(Some(client)).unwrap();
     let goalstate = provider.fetch_goalstate().unwrap();
 
@@ -244,5 +254,5 @@ fn test_goalstate_no_certs() {
 
     assert_eq!(goalstate.certs_endpoint(), None);
 
-    mockito::reset();
+    server.reset();
 }